Refactor stdin/file process logic into it's own package
